+++ test/src/common/plugins/plugins-config.spec.js
@@ -1,4 +1,5 @@
-import {evaluatePluginsConfig} from '../../../../src/common/plugins/plugins-config';
+import {evaluatePluginsConfig, getEncodedReferrer} from '../../../../src/common/plugins/plugins-config';
+let sandbox = sinon.sandbox.create();
 
 describe('evaluatePluginsConfig', function() {
   const playerConfig = {
@@ -35,5 +36,14 @@ describe('evaluatePluginsConfig', function() {
   it('should remove unevaluated plugins config from array', function() {
     evaluatePluginsConfig(pluginsConfig, playerConfig);
     pluginsConfig.kava.myArray.should.deep.equal([1, 'value', 0, true, false]);
+  });
+});
+
+describe('getEncodedReferrer', function() {
+  it('should encode the referrer', function() {
+    sandbox.stub(window.parent.document, 'URL').get(() => {
+      return 'http://localhost:3000/?debugKalturaPlayer';
+    });
+    getEncodedReferrer().should.be.equal('http%3A%2F%2Flocalhost%3A3000%2F%3FdebugKalturaPlayer');
   });
 });
