+++ test/src/common/utils/kaltura-params.spec.js
@@ -3,7 +3,6 @@ import {
   addKalturaParams,
   addReferrer,
   getReferrer,
-  getEncodedReferrer,
   handleSessionId,
   updateSessionIdInUrl
 } from '../../../../src/common/utils/kaltura-params';
@@ -15,7 +14,6 @@ class Player {
 }
 
 let player = new Player();
-let sandbox = sinon.sandbox.create();
 
 describe('addKalturaParams', function() {
   it('should add session id, referrer and client tag for playManifest source', function() {
@@ -140,15 +138,6 @@ describe('addReferrer', function() {
     player.config = {session: {}};
     addReferrer(source, player);
     source.url.should.be.equal('a/b/c/playmanifest/source?a&referrer=' + btoa(getReferrer().substr(0, 1000)));
-  });
-});
-
-describe('getEncodedReferrer', function() {
-  it('should encode the referrer', function() {
-    sandbox.stub(window.parent.document, 'URL').get(() => {
-      return 'http://localhost:3000/?debugKalturaPlayer';
-    });
-    getEncodedReferrer().should.be.equal('http%3A%2F%2Flocalhost%3A3000%2F%3FdebugKalturaPlayer');
   });
 });
